app_render.com/requirements.txt

gunicorn==23.0.0
Flask==2.3.2
google-cloud-firestore=2.20.1
